feat: 회원가입 탈퇴 사이드 메뉴에서 설정으로 이동
