+++ FlagBoard/FlagBoard/View/SideMenu/SideMenuView.storyboard
@@ -32,46 +32,39 @@
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="nJB-cf-zu8">
-                                <rect key="frame" x="164" y="88" width="86.5" height="35.5"/>
-                                <color key="tintColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <state key="normal" title="Button"/>
-                                <buttonConfiguration key="configuration" style="plain" title="회원가입">
-                                    <fontDescription key="titleFontDescription" type="system" pointSize="18"/>
-                                </buttonConfiguration>
-                                <connections>
-                                    <action selector="registerButtonPressed:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="PUk-QZ-oMn"/>
-                                </connections>
-                            </button>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="YXw-D5-1mu">
-                                <rect key="frame" x="171.5" y="144.5" width="71" height="35.5"/>
-                                <color key="tintColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <state key="normal" title="Button"/>
-                                <buttonConfiguration key="configuration" style="plain" title="로그인">
-                                    <fontDescription key="titleFontDescription" type="system" pointSize="18"/>
-                                </buttonConfiguration>
-                                <connections>
-                                    <action selector="didTabLoginButton:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="URn-FI-jqe"/>
-                                </connections>
-                            </button>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="xJE-a3-bSg">
-                                <rect key="frame" x="179.5" y="203" width="55.5" height="35.5"/>
-                                <color key="tintColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <state key="normal" title="Button"/>
-                                <buttonConfiguration key="configuration" style="plain" title="설정">
-                                    <fontDescription key="titleFontDescription" type="system" pointSize="18"/>
-                                </buttonConfiguration>
-                            </button>
+                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="15" translatesAutoresizingMaskIntoConstraints="NO" id="mZ2-LM-hPN">
+                                <rect key="frame" x="171.5" y="118" width="71" height="86"/>
+                                <subviews>
+                                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="YXw-D5-1mu">
+                                        <rect key="frame" x="0.0" y="0.0" width="71" height="35.5"/>
+                                        <color key="tintColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                        <state key="normal" title="Button"/>
+                                        <buttonConfiguration key="configuration" style="plain" title="로그인">
+                                            <fontDescription key="titleFontDescription" type="system" pointSize="18"/>
+                                        </buttonConfiguration>
+                                        <connections>
+                                            <action selector="loginButtonPressed:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="MeG-0u-Y2j"/>
+                                        </connections>
+                                    </button>
+                                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="pbK-3k-JIP">
+                                        <rect key="frame" x="0.0" y="50.5" width="71" height="35.5"/>
+                                        <color key="tintColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                        <state key="normal" title="Button"/>
+                                        <buttonConfiguration key="configuration" style="plain" title="설정">
+                                            <fontDescription key="titleFontDescription" type="system" pointSize="18"/>
+                                        </buttonConfiguration>
+                                        <connections>
+                                            <action selector="settingButtonPressed:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="VIo-eh-BDK"/>
+                                        </connections>
+                                    </button>
+                                </subviews>
+                            </stackView>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="vDu-zF-Fre"/>
                         <color key="backgroundColor" systemColor="opaqueSeparatorColor"/>
                         <constraints>
-                            <constraint firstItem="nJB-cf-zu8" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" id="7DJ-yO-9vO"/>
-                            <constraint firstItem="nJB-cf-zu8" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="IFZ-C5-33y"/>
-                            <constraint firstItem="YXw-D5-1mu" firstAttribute="top" secondItem="nJB-cf-zu8" secondAttribute="bottom" constant="21" id="U34-Ve-0tS"/>
-                            <constraint firstItem="YXw-D5-1mu" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="V1U-tl-dTY"/>
-                            <constraint firstItem="xJE-a3-bSg" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="XMC-Vf-SdL"/>
-                            <constraint firstItem="xJE-a3-bSg" firstAttribute="top" secondItem="YXw-D5-1mu" secondAttribute="bottom" constant="23" id="Za5-uO-uwR"/>
+                            <constraint firstItem="mZ2-LM-hPN" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="Gsb-nf-9Xz"/>
+                            <constraint firstItem="mZ2-LM-hPN" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" constant="30" id="nVW-2C-7pl"/>
                         </constraints>
                     </view>
                     <navigationItem key="navigationItem" id="USn-H8-59E"/>
